qal: do not update capture profile for pause/resume
When SVA stream 1 stops and capture profile needs to
be updated in rm, SVA stream 2 will do pause and
resume to update the capture profile. But in pause of
stream 2, it may call update capture profile to rm
again, as a result stream 1 pause will be called and
deadlock comes.

Do not update capture profile for pause/resume.

Change-Id: If70eb49b7689e3de87f6337bc5abb941c12f7c79