PAL: Update engine state after first detection in multi VA
In multi VA, after getting detection for one stream, engine
state is set to ENG_DETECTED, and while restarting recognition
for other active stream, engine state is checked, if it is
either active or buffring then only recognition is started
but engine state here will be detected, so recognition won't
be started for other active streams.

Hence we need to check state of other streams attached to same
engine and if there are other active streams we need to reset
engine's state to active.

Change-Id: I0a00f10924103d3f1bb73f99a0ed231428b7936f