pal: assign ring buffer reader based on sound model sequence
When readers created for stream and capi engines, we are assigning
with sequence assuming that GMM/PDK model always comes first in
sound model. But sometimes it may not be the first part of sound
model, which may cause null reader assigned to specifc capi engine.

Assign ring buffer reader based on sound model sequence.

Change-Id: Ib005986eeafc87028d2513fc83127dbbdb40d916