pal: Add fallback mechanism for custom config as selector
When a custom config is set and it doesn't need GKV pair
different than the default one, current implementation requires
it to be added in usecase manager XML to pick the GKVs.
Add a fallback mechanism to find the KVs by removing the
custom config as selector when the KV pairs are not
found with it.
Modify the way to get selector values for custom config using
device attributes.

Change-Id: I66c11b1ac92c04f65b3e6ee57e5b838f95f7402d