pal: force_switch condition for Play/Pause scnerio
No need to do force deviceswitch for all scenarios where device
is in stopped state. Need to check only if the device is suspended
and not started and try to perform a force device switch.

Otherwise there will be unnecessary device switches which are not
intended for. especially when 2 streams coming in and first stream
just connected started and second stream connect and tries to do
device switch of first stream.

Change-Id: Id9d0c70fc04dba29dffd6eff74e828a2cc5e9f25