pal: During ssr in error condition return read/write status
During ssr,  “errno == -ENETRESET &&  rm->cardState != CARD_STATUS_OFFLINE”
or “rm->cardState == CARD_STATUS_OFFLINE” Other than above conditions, in 
some cases errno is updated with positive value, which is sent to hal .In
hal it treats as success, so the outputs/inputs are not put into standby,
which is causing to undefined behavior. To avoid such issue, returns the
write/read status to HAL which is negative value and put the outputs/inputs
into standby.

Change-Id: Iac692b80e3f215a167e74bb04cbdf205d24ff10e