pal: Audio Leak related issues when a2dpSuspend
During a2dpSuspend - we move all the streams to speaker.
for offload streams we should be doing mute first and pause.
This is to ensure DSP has enough ramp down in Mute module.If
we issue pause first then no data processing happen and mute
ramp down will not happen which will then get processed after
resume , which will be percieved as a leak.

Changing logic to Mute first and pause later.

Change-Id: I552fab70802363e112a132cefb558f16efe15a70