pal: streamDevSwitch and StreamClose raceconditon.
DevSwitch usually operates on 2 steps.Disconnect of all
streams on that device and Connect of all streams on
a new device.

During that transistion we take respective streamMutex and
unlock after Disconnection, which paves the way for stream
operation on the stream there is no device connected e.g
stream close and other apis which will lead to failure due
to incorrect device count during stream destruction.

Fixing that with taking lock for the entier device switch
operation.

Change-Id: I3241f682a1566ca41a084b567368af15f113aeb6