pal: call processstartrecognition() for active streams only
If one voice ui stream in active state and other in loaded state,
any concurrency event will be posted to active stream and which leads
to active stream being unloaded and processstartrecognition() will be
called for other streams. As for loaded streams, mid_wakeup_cfg_ won't
be updated and it will be empty only, hence processstartrecognition()
will be called with null stream only. This is avoided with current
fix.

Change-Id: I5746e980c2c1fb06faaf118f32e3197ce73d2a08