pal: handle streams associated with combo devices during a2dpSuspend
Ringtone streams routed on combo devices(speaker and a2dp) is not well
handled during a2dp suspend scenario, because the same stream was
temporarily routed to speaker again and increase speaker device count
twice. As a result, speaker device is not closed.

To fix this, handle streams that associated with combo devices properly
during a2dp suspend and resume.

Change-Id: I794f32456f7c1aeb73fa8217f9e54696604e2378