pal: Handle concurrent non-tunnel stream instances
ResourceManager returns a numerically higher stream instance
when queried during stream creation for non-tunnel mode.
It does not account for whether the stream instance is available
at that point in time. Each stream instance maps to a module
instance in the graph and could be assigned to other concurrent
streams leading to a problem where the same graph is being used
by multiple streams.
Track active non-tunnel stream instances and assign an
available instance slot on opening a stream to avoid using
an already occupied graph.

Change-Id: Ibed4af0cfe30d1b702bc7bda75cc47e8722fa21a