pal: update the size check of activeStreamsDevices
In function disconnectSessionDevice, the size of
activeStreamsDevices is checked to determine if
device metadata needs to be freed. The current size
check could result the device metadata is never freed
for some streams. To avoid potential issues, update
this size check from 0 to 1 to free device metadata
when there is only one device left for disconnection.

Change-Id: I541fa1d70cf2b32a408f0d97e3884ead66ee85b7