PAL : Update conf level memory allocation check
If condition was only checking if memory is allocated
for conf levels or not, but in case of hotword number
of conf levels can be 0, which was causing issue, as
memory will not be allocated for this.

Now, checking if number of confidence level is greater than 0
and then if memory is not allocated then return error.

Change-Id: I41ccecf6991814355ff5ac49fd7f1324968a3d38