pal: Fix for dangling ec ref links on device switch
If end points have more than one reference to an ec link
and all streams running on that device are not closed. A
failure will occur when trying to open the new graph. This
fixes the issue by device switching all streams that are running
concurrently with the voice call stream that is device switching.

Change-Id: Ia06233df847a5a57c452e862e3c87194606e07af