PAL: A2DP to Speaker Switch Failures.
When A2DP switch Speaker/Handset - we should ensure the device
properities are correct.  if there are active streams on the device
query the current device properties and if no streams found use the
default configuration and sndDeviceName for the device to ensure
proper switch.

Scenario which this behavior happens is when there is
voice call happened and ended prior to A2DP UC. And when
A2DP suspend happens - will be moving to Speaker. In that workflow
sndDeviceId need to defaulted.

Change-Id: Ic2342defc663c3d141cdd992b8e872c1136961a1