pal: fix for no device switch occuring for shared backends
Device is not switching properly when the device configurations are the
same but the device ids are different. This change forces a device
switch if params are the same but device ids are different when there
is a shared backend. This in turn will casue the device switch to function
properly.

Change-Id: Ie631a4069d4e28f9e75694dc900987849197638f