pal: sequencing change for enabling external ec ref
Change for sequencing of external ec ref enablement for voice call.
External ec ref will now get enabled before the voice streams are setup.
This helps fix a pop issue that is seen during device switch. This is due
to a limitation in VCPM configuring the media format during a running state
that leads to a pop.

Change-Id: I49fb91b09363728c342e9f375cd4f9aa33d3c05d