pal: update correct return value in checkStreamMatch
we try to cross check stream_type and device_id received from hal
with stream/device attributes for a particular active stream using
checkStreamMatch. Wherein we check stream_type 1st and then device_id.
We have to return false if one of the condition is not met,
but we are not overwriting stream_type return value if device_id check is failed.
fixed the same by returning false if device_id not matched.

Without fix, set config fails continously for tx streams for
PARAM_ID_UI_EFFECT and the effects are not applying to rx stream
in case of concurrency with tx streams.

Change-Id: Ic0ae0055d06102817a1e61729ae025326d1ff4b2