pal: check rx stream/device activity before enable ec ref
When enable EC ref for Tx stream, EC ref count may be wrong
if there's any Rx stream which is stopped but yet to be
closed, and this may cause EC ref not disabled if other Rx
streams are stopped.

Check Rx stream/device activity when calculating ec ref count.

Change-Id: Ie98b13e28964e69a2ddfdae0202b7f2d846ba05c