PAL: Move Unmute/suspendclear after DeviceSwitch.
In some scenarios,when A2dpResume fail and Stream does not have
any device associated and in Subsequent DevSwitch tries to first
unmute the existing stream and go for the new incoming device switch.

This required proper caching with the AGM as the stream is not associated to any
device during this time. In-order to avoid perform the device unmute after the
incmoing device is connected.

Change-Id: Ifb3fca5d7b44f3d810c58be9efe03f88100c1b34