ar-pal: assign Idle cachedState in exit block in ssrUpHandler
Currently, the Idle cachedState assignment is outside of exit
block. If the cachedState were not assigned to be IDLE,
buffers would have been dropped and subroutine would falsely
regard SSR down/up handler was still in progress.

Even if stream open/start failed, we would still need to
assign cachedState to IDLE, so that proper write erros
can be sent to HAL, and retry opening/start stream.

Moving the assignment to exit block so that even if
stream open/start failed, Idle cachedState assignment
will be guaranteed to execute.

Change-Id: I28f4b61ac666f6682d388a1d59c363f0ce5e93a4