pal: Initialize customPayload and customPayloadSize
VSID set config failure is seen as the payload is getting
corrupted. This is because customPayload and customPayloadSize
are not initialized.
Set customPayload to NULL and customPayloadSize to 0 in
SessionAlsaVoice constructor to fix this issue.

Update the goto statement whenever setconfig of vsid fails
from exit to err_pcm_open

Change-Id: I687ec874223543134f49a8b60aeb4a2785dd29b1