pal: Notify concurrent stream status before stream start
While handling an LPI->NLPI switch for SoundTrigger streams from PAL,
concurrent stream handling currently takes place after the
respective stream is started. Change the sequence in such a way
that the SoundTrigger streams switch to NLPI before the stream
is started, we can avoid the memory peak in the transient
state of this switch and will not result in LPI memory exhaustion
as the SoundTrigger use case memory is freed up.

Change-Id: Ic04a14fd23a8162a356404b99c1371faf84ca7bb