pal: avoid voice/voip device switch if other stream use new device
During a voice or voip call, when a device switch happens for other
stream, then all streams sharing the same back end is re-routed to
the new device. This causes voice or voip call device switch happens
unexpectedly.

Fix this by using the sound devices of voice or voip call for other
stream if share backend.

Change-Id: Ic56a4482aaa294eba58fe003efc09fad119bb95a