pal: Fix SSR recovery in restore state mismatch scenario
In concurrent stream handling, state transitions occur from
active/loaded to idle and back to active/loaded. While
transitioning back to active/loaded, if SSR or some other
failure occurs in SPF, state for restore will remain as idle
and will not recover back to active or loaded in SSR online
handling. Add restore state check and set it to appropriate
states based on currentState of stream to fix this.

Change-Id: Ief6bf3ab57bdc9505d303852532a7d4daa01f9f7