pal: refine check to stop dual VA properly
LPI vote is not properly updated when voice call comes
during dualVA running with record concurrency disabled.
StopRecognition is executed because of voice call but
graph is not actually stopped for dualVA case then
when LPI/NLPI switch comes by some other playback stream,
graph is stopped and closed but lpi vote does not change.

Check if other stream is active to determine whether to
start graph in stop recognition.

Change-Id: I8815666368b67dbb5cc6db555cb7e9f4546751d5